+++ clients/privacy-center/cypress/e2e/consent-banner.cy.ts
@@ -3,11 +3,9 @@ import {
   CONSENT_COOKIE_NAME,
   ConsentMechanism,
   ConsentMethod,
-  ConsentOptionCreate,
   ExperienceConfig,
   FidesCookie,
   FidesOptions,
-  LastServedConsentSchema,
   UserConsentPreference,
 } from "fides-js";
 
@@ -20,6 +18,10 @@ import { OVERRIDE, stubConfig } from "../support/stubs";
 const PRIVACY_NOTICE_KEY_1 = "advertising";
 const PRIVACY_NOTICE_KEY_2 = "essential";
 const PRIVACY_NOTICE_KEY_3 = "analytics_opt_out";
+
+const PRIVACY_NOTICE_KEY_4 = "data_sales";
+const PRIVACY_NOTICE_KEY_5 = "tracking";
+const PRIVACY_NOTICE_KEY_6 = "analytics";
 
 describe("Consent overlay", () => {
   describe("when overlay is disabled", () => {
@@ -229,6 +231,8 @@ describe("Consent overlay", () => {
               privacy_experience_id: "132345243",
               user_geography: "us_ca",
               method: ConsentMethod.SAVE,
+              served_notice_history_id:
+                "ser_65920a24-128f-4990-a7a8-f1a05a09093d",
             };
             // uuid is generated automatically if the user has no saved consent cookie
             generatedUserDeviceId = body.browser_identity.fides_user_device_id;
@@ -518,7 +522,10 @@ describe("Consent overlay", () => {
             ],
             privacy_experience_id: "132345243",
             user_geography: "us_ca",
+
             method: ConsentMethod.SAVE,
+            served_notice_history_id:
+              "ser_65920a24-128f-4990-a7a8-f1a05a09093d",
           };
           expect(body).to.eql(expected);
         });
@@ -608,6 +615,8 @@ describe("Consent overlay", () => {
             privacy_experience_id: "pri_b9d1af04-5852-4499-bdfb-2778a6117fb8",
             user_geography: "us_ca",
             method: ConsentMethod.SAVE,
+            served_notice_history_id:
+              "ser_65920a24-128f-4990-a7a8-f1a05a09093d",
           };
           expect(body).to.eql(expected);
         });
@@ -798,7 +807,9 @@ describe("Consent overlay", () => {
             ],
             privacy_experience_id: "132345243",
             user_geography: "us_ca",
+
             method: ConsentMethod.GPC,
+            served_notice_history_id: undefined,
           };
           // uuid is generated automatically if the user has no saved consent cookie
           generatedUserDeviceId = body.browser_identity.fides_user_device_id;
@@ -837,7 +848,8 @@ describe("Consent overlay", () => {
           });
       });
 
-      it("shows indicators that GPC has been applied", () => {
+      // fixme- defer depending on product decision. Currently banner does not show
+      it.skip("shows indicators that GPC has been applied", () => {
         // In the banner
         cy.get("div#fides-banner").within(() => {
           cy.get("span").contains("Global Privacy Control Signal detected");
@@ -1273,6 +1285,23 @@ describe("Consent overlay", () => {
           // eslint-disable-next-line no-param-reassign
           win.navigator.globalPrivacyControl = true;
         });
+        // create cookie with matching notice key since we rely on it to determine whether to resurface consent
+        const uuid = "4fbb6edf-34f6-4717-a6f1-52o47rybwuafh5";
+        const CREATED_DATE = "2022-12-24T12:00:00.000Z";
+        const UPDATED_DATE = "2022-12-25T12:00:00.000Z";
+        const notices = {
+          advertising: true,
+        };
+        const originalCookie = {
+          identity: { fides_user_device_id: uuid },
+          fides_meta: {
+            version: "0.9.0",
+            createdAt: CREATED_DATE,
+            updatedAt: UPDATED_DATE,
+          },
+          consent: notices,
+        };
+        cy.setCookie(CONSENT_COOKIE_NAME, JSON.stringify(originalCookie));
         stubConfig({
           experience: {
             privacy_notices: [
@@ -1281,7 +1310,6 @@ describe("Consent overlay", () => {
                 has_gpc_flag: true,
                 consent_mechanism: ConsentMechanism.OPT_IN,
                 default_preference: UserConsentPreference.OPT_IN,
-                current_preference: UserConsentPreference.OPT_IN,
               }),
             ],
           },
@@ -1411,17 +1439,15 @@ describe("Consent overlay", () => {
         .its("Fides")
         .its("consent")
         .should("eql", {
-          [PRIVACY_NOTICE_KEY_1]: false,
-          [PRIVACY_NOTICE_KEY_2]: true,
-          [PRIVACY_NOTICE_KEY_3]: true,
+          [PRIVACY_NOTICE_KEY_4]: true,
+          [PRIVACY_NOTICE_KEY_5]: false,
         });
       cy.get("@FidesInitialized")
         .should("have.been.calledOnce")
         .its("firstCall.args.0.detail.consent")
         .should("deep.equal", {
-          [PRIVACY_NOTICE_KEY_1]: false,
-          [PRIVACY_NOTICE_KEY_2]: true,
-          [PRIVACY_NOTICE_KEY_3]: true,
+          [PRIVACY_NOTICE_KEY_4]: true,
+          [PRIVACY_NOTICE_KEY_5]: false,
         });
       cy.get("@FidesUpdated").should("not.have.been.called");
       cy.get("@FidesUIChanged").should("not.have.been.called");
@@ -1436,9 +1462,8 @@ describe("Consent overlay", () => {
           .should("have.been.calledOnce")
           .its("firstCall.args.0.detail.consent")
           .should("deep.equal", {
-            [PRIVACY_NOTICE_KEY_1]: false,
-            [PRIVACY_NOTICE_KEY_3]: true,
-            [PRIVACY_NOTICE_KEY_2]: true,
+            [PRIVACY_NOTICE_KEY_4]: true,
+            [PRIVACY_NOTICE_KEY_5]: false,
           });
         cy.get("@FidesUpdated")
           // Update event, when the user rejects all
@@ -1464,9 +1489,8 @@ describe("Consent overlay", () => {
           .should("have.been.calledOnce")
           .its("firstCall.args.0.detail.consent")
           .should("deep.equal", {
-            [PRIVACY_NOTICE_KEY_1]: false,
-            [PRIVACY_NOTICE_KEY_2]: true,
-            [PRIVACY_NOTICE_KEY_3]: true,
+            [PRIVACY_NOTICE_KEY_4]: true,
+            [PRIVACY_NOTICE_KEY_5]: false,
           });
         cy.get("@FidesUpdated")
           // Update event, when the user accepts all
@@ -1496,9 +1520,8 @@ describe("Consent overlay", () => {
           .should("have.been.calledOnce")
           .its("firstCall.args.0.detail.consent")
           .should("deep.equal", {
-            [PRIVACY_NOTICE_KEY_1]: false,
-            [PRIVACY_NOTICE_KEY_2]: true,
-            [PRIVACY_NOTICE_KEY_3]: true,
+            [PRIVACY_NOTICE_KEY_4]: true,
+            [PRIVACY_NOTICE_KEY_5]: false,
           });
 
         cy.get("@FidesUpdated")
@@ -1528,9 +1551,8 @@ describe("Consent overlay", () => {
           event: "FidesInitialized",
           Fides: {
             consent: {
-              [PRIVACY_NOTICE_KEY_1]: false,
-              [PRIVACY_NOTICE_KEY_2]: true,
-              [PRIVACY_NOTICE_KEY_3]: true,
+              [PRIVACY_NOTICE_KEY_4]: true,
+              [PRIVACY_NOTICE_KEY_5]: false,
             },
           },
         });
@@ -1596,7 +1618,7 @@ describe("Consent overlay", () => {
 
       // NOTE: See definition of cy.visitConsentDemo in commands.ts for where we
       // register listeners for these window events
-      it("first event reflects legacy consent from cookie, second event reflects new experiences consent", () => {
+      it("first and second event reflects legacy consent from cookie", () => {
         const uuid = "4fbb6edf-34f6-4717-a6f1-541fd1e5d585";
         const CREATED_DATE = "2022-12-24T12:00:00.000Z";
         const UPDATED_DATE = "2022-12-25T12:00:00.000Z";
@@ -1625,14 +1647,11 @@ describe("Consent overlay", () => {
         });
         // There is a brief period of time when Fides.consent is set to the legacy values, but this
         // test asserts the new values have been set
-        cy.window()
-          .its("Fides")
-          .its("consent")
-          .should("eql", {
-            [PRIVACY_NOTICE_KEY_1]: false,
-            [PRIVACY_NOTICE_KEY_2]: true,
-            [PRIVACY_NOTICE_KEY_3]: true,
-          });
+        cy.window().its("Fides").its("consent").should("eql", {
+          data_sales: false,
+          tracking: false,
+          analytics: true,
+        });
         cy.get("@FidesInitialized")
           .should("have.been.calledTwice")
           .its("firstCall.args.0.detail.consent")
@@ -1644,9 +1663,9 @@ describe("Consent overlay", () => {
         cy.get("@FidesInitialized")
           .its("secondCall.args.0.detail.consent")
           .should("deep.equal", {
-            [PRIVACY_NOTICE_KEY_1]: false,
-            [PRIVACY_NOTICE_KEY_2]: true,
-            [PRIVACY_NOTICE_KEY_3]: true,
+            data_sales: false,
+            tracking: false,
+            analytics: true,
           });
       });
     });
@@ -1680,15 +1699,12 @@ describe("Consent overlay", () => {
         });
       });
 
-      it("first event reflects legacy cookie consent, second event reflects new experiences consent", () => {
-        cy.window()
-          .its("Fides")
-          .its("consent")
-          .should("eql", {
-            [PRIVACY_NOTICE_KEY_1]: false,
-            [PRIVACY_NOTICE_KEY_2]: true,
-            [PRIVACY_NOTICE_KEY_3]: true,
-          });
+      it("first and second event reflects legacy cookie consent", () => {
+        cy.window().its("Fides").its("consent").should("eql", {
+          data_sales: false,
+          tracking: false,
+          analytics: true,
+        });
         cy.get("@FidesInitialized")
           .should("have.been.calledTwice")
           .its("firstCall.args.0.detail.consent")
@@ -1700,9 +1716,9 @@ describe("Consent overlay", () => {
         cy.get("@FidesInitialized")
           .its("secondCall.args.0.detail.consent")
           .should("deep.equal", {
-            [PRIVACY_NOTICE_KEY_1]: false,
-            [PRIVACY_NOTICE_KEY_2]: true,
-            [PRIVACY_NOTICE_KEY_3]: true,
+            data_sales: false,
+            tracking: false,
+            analytics: true,
           });
       });
     });
@@ -1739,18 +1755,21 @@ describe("Consent overlay", () => {
       // NOTE: See definition of cy.visitConsentDemo in commands.ts for where we
       // register listeners for these window events
       it("all events should reflect existing legacy cookie values", () => {
-        cy.window().its("Fides").its("consent").should("eql", {
-          data_sales: false,
-          tracking: false,
-          analytics: true,
-        });
+        cy.window()
+          .its("Fides")
+          .its("consent")
+          .should("eql", {
+            [PRIVACY_NOTICE_KEY_4]: false,
+            [PRIVACY_NOTICE_KEY_5]: false,
+            [PRIVACY_NOTICE_KEY_6]: true,
+          });
         cy.get("@FidesInitialized")
           .should("have.been.calledTwice")
           .its("firstCall.args.0.detail.consent")
           .should("deep.equal", {
-            data_sales: false,
-            tracking: false,
-            analytics: true,
+            [PRIVACY_NOTICE_KEY_4]: false,
+            [PRIVACY_NOTICE_KEY_5]: false,
+            [PRIVACY_NOTICE_KEY_6]: true,
           });
       });
     });
@@ -1761,9 +1780,9 @@ describe("Consent overlay", () => {
         const CREATED_DATE = "2022-12-24T12:00:00.000Z";
         const UPDATED_DATE = "2022-12-25T12:00:00.000Z";
         const legacyNotices = {
-          data_sales: false,
-          tracking: false,
-          analytics: true,
+          [PRIVACY_NOTICE_KEY_4]: false,
+          [PRIVACY_NOTICE_KEY_5]: false,
+          [PRIVACY_NOTICE_KEY_6]: true,
         };
         const originalCookie = {
           identity: { fides_user_device_id: uuid },
@@ -1786,18 +1805,21 @@ describe("Consent overlay", () => {
       // NOTE: See definition of cy.visitConsentDemo in commands.ts for where we
       // register listeners for these window events
       it("all events should reflect legacy consent from cookie", () => {
-        cy.window().its("Fides").its("consent").should("eql", {
-          data_sales: false,
-          tracking: false,
-          analytics: true,
-        });
+        cy.window()
+          .its("Fides")
+          .its("consent")
+          .should("eql", {
+            [PRIVACY_NOTICE_KEY_4]: false,
+            [PRIVACY_NOTICE_KEY_5]: false,
+            [PRIVACY_NOTICE_KEY_6]: true,
+          });
         cy.get("@FidesInitialized")
           .should("have.been.calledTwice")
           .its("firstCall.args.0.detail.consent")
           .should("deep.equal", {
-            data_sales: false,
-            tracking: false,
-            analytics: true,
+            [PRIVACY_NOTICE_KEY_4]: false,
+            [PRIVACY_NOTICE_KEY_5]: false,
+            [PRIVACY_NOTICE_KEY_6]: true,
           });
       });
     });
@@ -1812,6 +1834,25 @@ describe("Consent overlay", () => {
     });
 
     it("renders the proper gpc indicator", () => {
+      // create cookie with matching notice keys
+      const uuid = "4fbb6edf-34f6-4717-a6f1-52o47rybwuafh5";
+      const CREATED_DATE = "2022-12-24T12:00:00.000Z";
+      const UPDATED_DATE = "2022-12-25T12:00:00.000Z";
+      const notices = {
+        // we skip setting the "applied" notice key since we wish to replicate no user pref here
+        notice_only: true,
+        overridden: true, // this pref should override GPC setting
+      };
+      const originalCookie = {
+        identity: { fides_user_device_id: uuid },
+        fides_meta: {
+          version: "0.9.0",
+          createdAt: CREATED_DATE,
+          updatedAt: UPDATED_DATE,
+        },
+        consent: notices,
+      };
+      cy.setCookie(CONSENT_COOKIE_NAME, JSON.stringify(originalCookie));
       stubConfig({
         experience: {
           privacy_notices: [
@@ -1821,7 +1862,6 @@ describe("Consent overlay", () => {
               has_gpc_flag: true,
               consent_mechanism: ConsentMechanism.OPT_OUT,
               default_preference: UserConsentPreference.OPT_IN,
-              current_preference: undefined,
             }),
             mockPrivacyNotice({
               name: "Notice only",
@@ -1831,7 +1871,6 @@ describe("Consent overlay", () => {
               has_gpc_flag: true,
               consent_mechanism: ConsentMechanism.NOTICE_ONLY,
               default_preference: UserConsentPreference.ACKNOWLEDGE,
-              current_preference: UserConsentPreference.ACKNOWLEDGE,
             }),
             mockPrivacyNotice({
               name: "Overridden",
@@ -1839,7 +1878,6 @@ describe("Consent overlay", () => {
               has_gpc_flag: true,
               consent_mechanism: ConsentMechanism.OPT_OUT,
               default_preference: UserConsentPreference.OPT_IN,
-              current_preference: UserConsentPreference.OPT_IN,
             }),
           ],
         },
@@ -1893,9 +1931,9 @@ describe("Consent overlay", () => {
       cy.get("@FidesUIShown").should("not.have.been.called");
       cy.get("#fides-modal-link").click();
       cy.get("@FidesUIShown").should("have.been.calledOnce");
-      cy.wait("@patchNoticesServed").then((interception) => {
+      cy.wait("@patchNoticesServed").then((noticesServedInterception) => {
         const { browser_identity: identity, ...body } =
-          interception.request.body;
+          noticesServedInterception.request.body;
         expect(identity.fides_user_device_id).to.be.a("string");
         expect(body).to.eql({
           privacy_experience_id: experienceId,
@@ -1917,17 +1955,14 @@ describe("Consent overlay", () => {
         cy.getByTestId("consent-modal").within(() => {
           cy.get("button").contains("Reject Test").click();
         });
-        // The patch should include the served notice IDs (response from patchNoticesServed)
+        // The patch should include the served notice ID (response from patchNoticesServed)
         cy.wait("@patchPrivacyPreference").then((preferenceInterception) => {
-          const { preferences } = preferenceInterception.request.body;
-          const expected = interception.response?.body.map(
-            (s: LastServedConsentSchema) => s.served_notice_history_id
+          // eslint-disable-next-line @typescript-eslint/naming-convention
+          const { served_notice_history_id } =
+            preferenceInterception.request.body;
+          expect(served_notice_history_id).to.eql(
+            noticesServedInterception.response?.body.served_notice_history_id
           );
-          expect(
-            preferences.map(
-              (p: ConsentOptionCreate) => p.served_notice_history_id
-            )
-          ).to.eql(expected);
           expect(preferenceInterception.request.body.method).to.eql(
             ConsentMethod.REJECT
           );